+++ migrations/005_create_referrals.down.sql
@@ -1,0 +1,2 @@
+DROP TABLE IF EXISTS user_referral_codes;
+DROP TABLE IF EXISTS referrals;